rename modules to drop wal prefix
