+++ slatedb/src/wal/slatedb/mod.rs
@@ -2,8 +2,8 @@
 
 pub(crate) mod admin;
 pub(crate) mod gc;
+pub(crate) mod iterator;
 pub(crate) mod reader;
-pub(crate) mod wal_iterator;
-pub(crate) mod wal_sst_builder;
-pub(crate) mod wal_writer;
+pub(crate) mod sst_builder;
+pub(crate) mod writer;
 pub(crate) mod writer_init;
