+++ slatedb/src/wal/slatedb/writer_init.rs
@@ -5,8 +5,8 @@ use crate::manifest::Manifest;
 use crate::sst_iter::SstIteratorOptions;
 use crate::tablestore::TableStore;
 use crate::utils::WatchableOnceCellReader;
-use crate::wal::slatedb::wal_iterator::{SlateDbWalIterator, SlateDbWalIteratorOptions};
-use crate::wal::slatedb::wal_writer::SlateDbWalWriter;
+use crate::wal::slatedb::iterator::{SlateDbWalIterator, SlateDbWalIteratorOptions};
+use crate::wal::slatedb::writer::SlateDbWalWriter;
 use crate::wal::{WalError, WriterInitResult, WriterManifest};
 use crate::{wal, Settings};
 use async_trait::async_trait;
